% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code.R
\name{get_code}
\alias{get_code}
\title{Retrieve state and district codes}
\usage{
get_code(regions = c("states", "state", "districts", "district"))
}
\arguments{
\item{regions}{The region breakdown for the map, can be one of
(\code{"states"}, \code{"state"}, \code{"districts"}, \code{"district"}).
The default is \code{"states"}.}
}
\value{
An data frame of codes of the desired \code{regions}.
}
\description{
Retrieve state and district codes
}
\examples{
str(get_code())

state_codes <- get_code()
county_codes <- get_code(regions = "districts")

}
