% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_code11.R
\name{get_code11}
\alias{get_code11}
\title{Retrieve state and district codes}
\usage{
get_code11(regions = c("states", "state", "districts", "district"))
}
\arguments{
\item{regions}{The region breakdown for the map, can be one of
(\code{"states"}, \code{"state"}, \code{"districts"}, \code{"district"}).
The default is \code{"states"}.}
}
\value{
An data frame of codes of the desired \code{regions}.
}
\description{
Retrieve state and district codes
}
\examples{
str(get_code11())

state_codes <- get_code11()
county_codes <- get_code11(regions = "districts")

}
