% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiamapdata-package.R
\docType{package}
\name{indiamapdata-package}
\alias{indiamapdata}
\alias{indiamapdata-package}
\title{indiamapdata: Mapping Data for 'indiamap' Package}
\description{
Provides a container for data used by the 'indiamap' package. The data used by 'indiamap' has been extracted into this package so that the file size of the 'indiamap' package can be reduced considerably. The data in this package will be updated when latest data is available.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/shubhamdutta26/indiamapdata}
  \item Report bugs at \url{https://github.com/shubhamdutta26/indiamapdata/issues}
}

}
\author{
\strong{Maintainer}: Shubham Dutta \email{shubhamdutta26@gmail.com} (\href{https://orcid.org/0000-0001-8484-0717}{ORCID}) [copyright holder]

}
\keyword{internal}
