% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/india_map.R
\name{india_map}
\alias{india_map}
\title{Xxx}
\usage{
india_map(
  regions = c("states", "state", "districts", "district"),
  include = c(),
  exclude = c()
)
}
\arguments{
\item{regions}{xxx}

\item{include}{xxx}

\item{exclude}{xxx}
}
\value{
xxx
}
\description{
Xxx
}
